Split KResource into a file. Include an interface.
